Extract Movie struct and AssertContains to files
